+++ apps/endringsmelding-pleiepenger/src/app/søknad/steps/oppsummering/OppsummeringStep.tsx
@@ -72,9 +72,7 @@ const OppsummeringStep = () => {
     const harGyldigArbeidstid = oppsummeringStepUtils.erArbeidstidEndringerGyldig(arbeidstid);
     const lovbestemtFerieErEndret = oppsummeringStepUtils.harEndringerILovbestemtFerieApiData(lovbestemtFerie);
 
-    const harIngenEndringer =
-        (valgteEndringer.arbeidstid && arbeidstidErEndret === false) ||
-        (valgteEndringer.lovbestemtFerie && lovbestemtFerieErEndret === false);
+    const harIngenEndringer = arbeidstidErEndret === false && lovbestemtFerieErEndret === false;
 
     return (
         <SøknadStep stepId={stepId} stepConfig={stepConfig}>
@@ -87,10 +85,10 @@ const OppsummeringStep = () => {
                 </Ingress>
             </SifGuidePanel>
 
-            {sak.harUkjentArbeidsforhold && ukjenteArbeidsforhold && (
+            {sak.harArbeidsgivereIkkeISak && ukjenteArbeidsforhold && (
                 <Block margin="xxl">
                     <SummarySection header="Nytt arbeidsforhold">
-                        {sak.ukjenteArbeidsgivere.map((arbeidsgiver) => {
+                        {sak.arbeidsgivereIkkeISak.map((arbeidsgiver) => {
                             const arbeidsforhold = ukjenteArbeidsforhold.find(
                                 (a) => a.organisasjonsnummer === arbeidsgiver.organisasjonsnummer,
                             );
@@ -145,7 +143,7 @@ const OppsummeringStep = () => {
                             <>
                                 <ArbeidstidOppsummering
                                     arbeidstid={arbeidstid}
-                                    arbeidsgivere={[...arbeidsgivere, ...sak.ukjenteArbeidsgivere]}
+                                    arbeidsgivere={[...arbeidsgivere, ...sak.arbeidsgivereIkkeISak]}
                                 />
                                 {!harGyldigArbeidstid && (
                                     <Block margin="none" padBottom="l">