+++ apps/pleiepenger-sykt-barn/src/app/types/Søkerdata.ts
@@ -1,0 +1,7 @@
+import { Søker } from '.';
+import { RegistrerteBarn } from './RegistrerteBarn';
+
+export interface Søkerdata {
+    søker: Søker;
+    barn: RegistrerteBarn[];
+}